+++ storage-node/src/services/caching/localDataObjects.ts
@@ -1,11 +1,14 @@
 import AwaitLock from 'await-lock'
-import path from 'path'
 import fs from 'fs'
+import path from 'path'
 import logger from '../logger'
 const fsPromises = fs.promises
 
+type DataObjectId = string
+type DataObjectPinCount = number
+
 // Local in-memory cache for IDs.
-let idCache = new Set<string>()
+const idCache = new Map<DataObjectId, DataObjectPinCount>()
 
 const lock = new AwaitLock()
 
@@ -17,7 +20,7 @@ const lock = new AwaitLock()
  */
 export async function getDataObjectIDs(): Promise<string[]> {
   await lock.acquireAsync()
-  const ids = Array.from(idCache)
+  const ids = [...idCache.keys()]
   lock.release()
 
   return ids
@@ -46,8 +49,7 @@ export async function loadDataObjectIdCache(
     (dataObjectId) => dataObjectId !== tempDirectoryName && dataObjectId !== pendingDirectoryName
   )
 
-  idCache = new Set(ids)
-
+  ids.forEach((id) => idCache.set(id, 0))
   logger.debug(`Local ID cache loaded.`)
 
   lock.release()
@@ -56,22 +58,50 @@ export async function loadDataObjectIdCache(
 /**
  * Adds data object ID to the local cache.
  *
- * @param dataObjectId - uploading directory
+ * @param dataObjectId - Storage data object ID
  *
  * @returns empty promise.
  */
 export async function addDataObjectIdToCache(dataObjectId: string): Promise<void> {
   await lock.acquireAsync()
 
-  idCache.add(dataObjectId)
+  idCache.set(dataObjectId, 0)
 
+  lock.release()
+}
+
+/**
+ * Pins data object ID in the local cache (increments the data object usage).
+ *
+ * @param dataObjectId - Storage data object ID
+ *
+ * @returns empty promise.
+ */
+export async function pinDataObjectIdToCache(dataObjectId: string): Promise<void> {
+  await lock.acquireAsync()
+  const currentPinnedCount = idCache.get(dataObjectId) || 0
+  idCache.set(dataObjectId, currentPinnedCount + 1)
+  lock.release()
+}
+
+/**
+ * Un-pins data object ID from the local cache (decrements the data object usage).
+ *
+ * @param dataObjectId - Storage data object ID
+ *
+ * @returns empty promise.
+ */
+export async function unpinDataObjectIdFromCache(dataObjectId: string): Promise<void> {
+  await lock.acquireAsync()
+  const currentPinnedCount = idCache.get(dataObjectId)
+  idCache.set(dataObjectId, currentPinnedCount ? currentPinnedCount - 1 : 0)
   lock.release()
 }
 
 /**
  * Deletes data object ID from the local cache.
  *
- * @param dataObjectId - uploading directory
+ * @param dataObjectId - Storage data object ID
  */
 export async function deleteDataObjectIdFromCache(dataObjectId: string): Promise<void> {
   await lock.acquireAsync()
@@ -79,6 +109,26 @@ export async function deleteDataObjectIdFromCache(dataObjectId: string): Promise
   idCache.delete(dataObjectId)
 
   lock.release()
+}
+
+/**
+ * Get data object ID from the local cache.
+ *
+ * @param dataObjectId - Storage data object ID
+ */
+export async function getDataObjectIdFromCache(
+  dataObjectId: string
+): Promise<{ dataObjectId: DataObjectId; pinnedCount: DataObjectPinCount } | undefined> {
+  // First check without lock
+  if (!idCache.has(dataObjectId)) {
+    return undefined
+  }
+
+  // Acquire lock
+  await lock.acquireAsync()
+  const id = { dataObjectId, pinnedCount: idCache.get(dataObjectId) as DataObjectPinCount }
+  lock.release()
+  return id
 }
 
 /**